third_party/blink/web_tests/fast/dom/HTMLImageElement/image-natural-width-height-svg-expected.txt

This is a testharness.js-based test.
[FAIL] naturalWidth/Height of SVG in <img>, width in pixels; height unspecified
  assert_equals: expected "500x0" but got "500x150"
[FAIL] naturalWidth/Height of SVG in <img>, width in pixels; percentage height
  assert_equals: expected "500x0" but got "500x150"
[FAIL] naturalWidth/Height of SVG in <img>, width/height unspecified; viewBox
  assert_equals: expected "0x0" but got "200x150"
[FAIL] naturalWidth/Height of SVG in <img>, width in pixels; height unspecified; viewBox
  assert_equals: expected "400x0" but got "400x300"
Harness: the test ran to completion.
